requirements.txt
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
rich>=13.0.0
